Day14 - My brother recommended I skip to this one
Definitely an interesting challenge, the solution to part one was very
easy and quick.  The issue with part two took longer than I like as I
was looking for the robots to ALL be adjacent to each other.  The check
to ensure there was no overlaps is not the best solution here either as the
ones that are not part of the tree could overlap.  Not happy with this
solution but time really is a premium outside of work.